+++ newm-chain/src/main/kotlin/io/newm/chain/daemon/DaemonsKoinModule.kt
@@ -4,5 +4,5 @@ import org.koin.dsl.bind
 import org.koin.dsl.module
 
 val daemonsKoinModule = module {
-    single { BlockDaemon(get()) } bind Daemon::class
+    single { BlockDaemon(get(), get(), get()) } bind Daemon::class
 }
