+++ newm-chain/src/main/kotlin/io/newm/chain/di/DependencyInjectionInstall.kt
@@ -1,7 +1,9 @@
 package io.newm.chain.di
 
-import io.ktor.server.application.*
+import io.ktor.server.application.Application
+import io.ktor.server.application.install
 import io.newm.chain.daemon.daemonsKoinModule
+import io.newm.chain.database.databaseKoinModule
 import org.koin.dsl.module
 import org.koin.ktor.plugin.Koin
 import org.slf4j.LoggerFactory
@@ -16,6 +18,7 @@ fun Application.installDependencyInjection() {
         modules(
             appKoinModule,
             daemonsKoinModule,
+            databaseKoinModule,
         )
     }
 }
